feat: support recording with no prompt
This change adds a new hotkey that supports transcription with no prompt.

This ends up being 'faster' and doesn't use any prompting before copy/pasting.

The concept of recording modes also is added to better distinguish between:
  - 'with default prompt'
  - 'with no prompt'
  - 'with prompt selection'

Renamed the 'direct' in code to be 'default' in terms of the modes.